backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pandas
python-multipart
pypdf2
python-docx
langchain-google-genai